refactor: split functions to Layered Architecture with Modular Monolith in Electron Main Process
