+++ src/main/database/connection.ts
@@ -1,6 +1,5 @@
 import path from 'node:path'
-
-import { app, ipcMain } from 'electron'
+import { app } from 'electron'
 
 import type { SqliteDatabase } from './types'
 
@@ -9,10 +8,10 @@ let db: SqliteDatabase | undefined
 function createDatabase(dbPath: string): SqliteDatabase {
   try {
     const BetterSqlite3 = require('better-sqlite3') as typeof import('better-sqlite3')
-
     return new BetterSqlite3(dbPath)
   } catch (error) {
     const errorMessage = error instanceof Error ? error.message : String(error)
+
     console.warn(`better-sqlite3 nicht verfuegbar (${errorMessage}). Nutze node:sqlite Fallback.`)
 
     const { DatabaseSync } = require('node:sqlite') as typeof import('node:sqlite')
@@ -29,22 +28,10 @@ export function initDatabase(): SqliteDatabase {
   return db
 }
 
-ipcMain.handle('get-users', () => {
-  return db!.prepare('SELECT * FROM users').all()
-})
-
-ipcMain.handle('add-user', (_event, user: { name: string; data: unknown }) => {
-  const insert = db!.prepare('INSERT INTO users (name, data) VALUES (?, ?)')
-  return insert.run(user.name, JSON.stringify(user.data))
-})
-
-ipcMain.handle('db-healthcheck', () => {
-  const result = db!.prepare('SELECT sqlite_version() AS version').get() as
-    | { version: string }
-    | undefined
-
-  return {
-    ok: true,
-    version: result?.version ?? 'unknown'
+export function getDatabase(): SqliteDatabase {
+  if (!db) {
+    throw new Error('Database wurde noch nicht initialisiert.')
   }
-})
+
+  return db
+}
