+++ src/main/ipc/register.ts
@@ -1,0 +1,7 @@
+import { registerHealthIpc } from './health'
+import { registerUsersIpc } from './users'
+
+export function registerIpcHandlers() {
+  registerHealthIpc()
+  registerUsersIpc()
+}
